Feature ETP-4708: Extract DetailView's module-scope helpers (T31)
DetailView.jsx: 5131 -> 3563 lines. 107 module-scope functions moved to a
sibling detailViewHelpers.jsx: pure helpers, predicates, className/style getters
and presentational fragments that took explicit params and held no component
state. They lived in the god component only because that is where they grew.

Behaviour is unchanged and no test was edited (R1). DetailView.jsx re-exports the
76 names its suites import from '../DetailView.jsx', so every existing import
keeps resolving to the same function; only the definition site moved.

The new file is added to the file-lines ratchet in THIS commit, before lowering
DetailView's baseline. Tracking the destination is what stops an extraction from
reading as payment: without it the 1.568 relocated lines would show up as a win
while the code simply moved somewhere unwatched — the same false zero the leak
scanner produced when the tab strip moved to TabStrip.jsx (churn report §9.3).

Verified: contract-ui vitest 2520/2520 (identical to the pre-move baseline),
3 E2E golden masters 36/36, window-leak ratchet 4 at baseline 4.
